+++ src/main/java/org/seec/muggle/auror/vo/strategy/event/EventVO.java
@@ -22,7 +22,7 @@ public class EventVO {
     String couponName;
     String couponDescription;
     String couponPictureUrl;
-    Double couponDiscount;
+    Integer couponDiscount;
     Integer couponThreshold;
     String couponExpiration;
 
@@ -99,11 +99,11 @@ public class EventVO {
         this.couponPictureUrl = couponPictureUrl;
     }
 
-    public Double getCouponDiscount() {
+    public Integer getCouponDiscount() {
         return couponDiscount;
     }
 
-    public void setCouponDiscount(Double couponDiscount) {
+    public void setCouponDiscount(Integer couponDiscount) {
         this.couponDiscount = couponDiscount;
     }
